requirements.txt
Django==4.2.8
djangorestframework==3.14.0
psycopg2-binary==2.9.9
Pillow==10.1.0
djoser==2.2.2
djangorestframework-simplejwt==5.3.1
boto3==1.33.11
django-storages==1.14.2
razorpay==1.4.1
coreapi==2.3.3